+++ spec/cucumber/formatter/html_spec.rb
@@ -97,6 +97,17 @@ module Cucumber
           end
         end
 
+        describe "with a comment at scenario level" do
+          define_feature <<-FEATURE
+            Feature: Foo
+              # Healthy Scenario
+              Scenario:
+                Given passing
+          FEATURE
+
+          it { expect(@doc).to have_css_node('.scenario .comment', /Healthy Scenario/) }
+        end
+
         describe "with a tag" do
           define_feature <<-FEATURE
             @foo