+++ History.md
@@ -9,6 +9,7 @@
 
 ### Bugfixes
 
+* Pass tags and comments to the legacy api formatters correctly, fixes [864](https://github.com/cucumber/cucumber/pull/864) (@brasmusson).
 * Properly close HTML formatter pre tag for leading comment ([828](https://github.com/cucumber/cucumber/issues/828), [854](https://github.com/cucumber/cucumber/pull/854) @pdswan)
 * Make the HTML formatter handle failures in hooks ([835](https://github.com/cucumber/cucumber/issues/835) @brasmusson)
 * Fix the check for multiple formatters using stdout ([826](https://github.com/cucumber/cucumber/pull/826) @brasmusson)